+++ wrappers/Mathcad/src/RPstrings.cpp
@@ -53,10 +53,10 @@ std::string format(const char* fmt, ...)
     std::shared_ptr<char> buffer(new char[size], deleter::delarray); // I'd prefer unique_ptr, but it's only available since c++11
     va_list vl;
     va_start(vl,fmt);
-    int nsize = vsnprintf(buffer.get(),size,fmt,vl);
-    if(size<=nsize){//fail delete buffer and try again
-        buffer.reset(new char[++nsize], deleter::delarray);//+1 for /0
-        nsize = vsnprintf(buffer.get(),nsize,fmt,vl);
+    int nsize = vsnprintf(buffer.get(),(size_t)size,fmt,vl);
+    if(size<=nsize){                                         //fail delete buffer and try again
+        buffer.reset(new char[++nsize], deleter::delarray);    //+1 for /0
+        nsize = vsnprintf(buffer.get(),(size_t)nsize,fmt,vl);
     }
     va_end(vl);
     return buffer.get();
